+++ desktop/qml/LocationsPage.qml
@@ -31,6 +31,7 @@ Page {
     property alias season: seasonSpinBox.season
     property alias hasSelection: locationView.hasSelection
     property alias rowCount: locationView.rowCount
+    property bool showPlantingsPane: true
 
     function refresh() {
         locationView.refresh();
@@ -55,6 +56,14 @@ Page {
         enabled: navigationIndex === 2 && filterField.visible && !addDialog.activeFocus && !editDialog.activeFocus
         context: Qt.ApplicationShortcut
         onActivated: filterField.forceActiveFocus();
+    }
+
+    Shortcut {
+        sequence: "Ctrl+P"
+        enabled: navigationIndex === 2 && filterField.visible && !addDialog.activeFocus && !editDialog.activeFocus
+
+        context: Qt.ApplicationShortcut
+        onActivated: showPlantingPaneButton.clicked()
     }
 
     Shortcut {
@@ -95,6 +104,41 @@ Page {
         enabled: navigationIndex === 2 && filterField.visible && !addDialog.activeFocus && !editDialog.activeFocus
         context: Qt.ApplicationShortcut
         onActivated: seasonSpinBox.previousYear();
+    }
+
+    Shortcut {
+        sequence: "Shift+A"
+        enabled: navigationIndex === 2 && filterField.visible && !addDialog.activeFocus && !editDialog.activeFocus
+        context: Qt.ApplicationShortcut
+        onActivated: expandButton.expandLevel(0)
+    }
+
+    Shortcut {
+        sequence: "Shift+B"
+        enabled: navigationIndex === 2 && filterField.visible && !addDialog.activeFocus && !editDialog.activeFocus
+        context: Qt.ApplicationShortcut
+        onActivated: expandButton.expandLevel(1)
+    }
+
+    Shortcut {
+        sequence: "Shift+C"
+        enabled: navigationIndex === 2 && filterField.visible && !addDialog.activeFocus && !editDialog.activeFocus
+        context: Qt.ApplicationShortcut
+        onActivated: expandButton.expandLevel(2)
+    }
+
+    Shortcut {
+        sequence: "Shift+D"
+        enabled: navigationIndex === 2 && filterField.visible && !addDialog.activeFocus && !editDialog.activeFocus
+        context: Qt.ApplicationShortcut
+        onActivated: expandButton.expandLevel(3)
+    }
+
+    Shortcut {
+        sequence: "Shift+E"
+        enabled: navigationIndex === 2 && filterField.visible && !addDialog.activeFocus && !editDialog.activeFocus
+        context: Qt.ApplicationShortcut
+        onActivated: expandButton.expandLevel(4)
     }
 
     onEditModeChanged: {
@@ -225,6 +269,56 @@ Page {
                 }
             }
 
+            RoundButton {
+                id: expandButton
+                flat: true
+                text: "\ue313"
+                font.family: "Material Icons"
+                font.pixelSize: 24
+                onClicked: expandMenu.open();
+
+                property int depth: locationView.treeDepth
+                property var expandBoolList: ({})
+
+                function expandLevel(level) {
+                    if (expandBoolList[level]) {
+                        locationView.collapseAll(level, false)
+                    } else {
+                        locationView.expandAll(level)
+                        for (var i = 0; i < level; i++)
+                            expandBoolList[i] = true
+                    }
+                    expandBoolList[level] = !expandBoolList[level]
+                    expandBoolListChanged();
+                }
+
+                Component.onCompleted: {
+                    for (var i; i < depth; i++)
+                        expandBoolList[i+1] = false
+                }
+
+                ToolTip.text: qsTr("Expand and collapse location levels")
+                ToolTip.delay: Units.shortDuration
+                ToolTip.visible: hovered
+
+                Menu {
+                    id: expandMenu
+                    y: expandButton.height
+
+                    Repeater {
+                        model: expandButton.depth
+                        MenuItem {
+                            text: index + 1
+                            checkable: true
+                            checked: expandButton.expandBoolList[index]
+                                     ? expandButton.expandBoolList[index]
+                                     : false
+                            onTriggered: expandButton.expandLevel(index)
+                        }
+                    }
+                }
+            }
+
             Button {
                 id: duplicateButton
                 flat: true
@@ -261,13 +355,13 @@ Page {
                 }
             }
 
-            CheckBox {
-                id: unassignedPlantingsCheckbox
-                text: qsTr("Show unassigned plantings")
-                Layout.leftMargin: 16
-                visible: !editMode
-                checked: true
-            }
+//            CheckBox {
+//                id: unassignedPlantingsCheckbox
+//                text: qsTr("Show unassigned plantings")
+//                Layout.leftMargin: 16
+//                visible: !editMode
+//                checked: true
+//            }
 
             CheckBox {
                 id: emptyLocationsCheckbox
@@ -313,7 +407,7 @@ Page {
             top: topDivider.bottom
             bottom: parent.bottom
             horizontalCenter: parent.horizontalCenter
-            margins: Units.smallSpacing
+            margins: 0
         }
         spacing: Units.smallSpacing
         width: plantingsView.implicitWidth
@@ -375,19 +469,65 @@ Page {
 
         Pane {
             id: plantingsPane
-            visible: unassignedPlantingsCheckbox.checked & !editMode
+//            visible: unassignedPlantingsCheckbox.checked & !editMode
+            visible: !editMode
+
             padding: 0
             Layout.fillWidth: true
             //            Layout.fillHeight: true
-            Layout.minimumHeight: page.height / 4
+            Layout.minimumHeight: showPlantingsPane ? page.height / 4 : 0
+//            Layout.minimumHeight: unassignedPlantingsCheckbox.checked ? page.height / 4 : 10
             Material.elevation: 2
             Material.background: "white"
 
+            Behavior on Layout.minimumHeight {
+                NumberAnimation { duration: Units.shortDuration  }
+            }
+
+            MouseArea {
+                id: plantingPaneMouseArea
+                anchors.fill: parent
+                hoverEnabled: true
+
+            RoundButton {
+                id: showPlantingPaneButton
+                z: -1
+                Material.background: "white"
+                width: 60
+                height: width
+                anchors.top: parent.top
+                anchors.topMargin: visible ? -width/2 : 0
+                anchors.horizontalCenter: parent.horizontalCenter
+                onClicked: showPlantingsPane =!showPlantingsPane
+                contentItem: Text {
+                    text: showPlantingsPane ? "\ue313" : "\ue316"
+                    font.family: "Material Icons"
+                    font.pixelSize: 24
+//                    color: parent.down ? "#17a81a" : "#21be2b"
+                    horizontalAlignment: Text.AlignHCenter
+                    verticalAlignment: Text.Top
+                    elide: Text.ElideRight
+                }
+
+                ToolTip.visible: hovered
+                ToolTip.text: showPlantingsPane ? qsTr("Hide the plantings pane") : qsTr("Show the planting pane")
+                ToolTip.delay: Units.shortDuration
+
+                Behavior on anchors.topMargin {
+                    NumberAnimation { duration: Units.shortDuration  }
+                }
+            }
+
+            Rectangle {
+                color: "white"
+                anchors.fill: parent
+            }
+
             Column {
                 id: emptyPlantingStateColumn
-                z: 1
                 spacing: Units.smallSpacing
-                visible: !plantingsView.rowsNumber
+                visible: !plantingsView.rowsNumber && showPlantingsPane
+                z:2
                 anchors {
                     centerIn: parent
                 }
@@ -484,3 +624,4 @@ Page {
     }
 }
 
+}